massive reorganization; still needs cleanup
