+++ patches/__init__.py
@@ -2,4 +2,4 @@ import patches.ips
 
 from .compression import lz77_decompress, lzma_compress
 from .firmware import Device, ExtFirmware, Firmware, IntFirmware
-from .patches import add_patch_args, apply_patches, validate_patch_args
+from .mario import MarioGnW
